+++ HISTORY.rst
@@ -162,6 +162,7 @@
 - Update Docker Compose volume names in documentation.
 - Split the Docker Compose installation chapter into multiple files.
 - Remove `importlib-metadata`.
+- Expose Django's `CACHES` setting as `MAYAN_CACHES`.
 
 4.9.6 (XXXX-XX-XX)
 ==================